fix(incremental): PURGE_WEIGHT is i64 — -(1 << 60) does not fit a 32-bit isize
`cargo check --target wasm32-unknown-unknown` failed with
  error[E0080]: attempt to shift left by `60_i32`, which would overflow
   --> core/incremental/compiler.rs:1444:50
because isize is 32-bit on wasm32, so the recursive-state purge sentinel
could not be evaluated at all.

The stored weight column is a SQLite integer, so i64 is the type that
actually describes it. WriteRow's weight path in persistence.rs now carries
i64 end to end — which also drops the `w as isize` narrowing that silently
truncated stored weights on 32-bit targets. Operator call sites that hold an
isize weight cast at the boundary.

+++ core/incremental/antijoin_operator.rs
@@ -774,7 +774,7 @@ impl IncrementalOperator for AntijoinOperator {
                     return_and_restore_if_io!(
                         &mut self.commit_state,
                         state,
-                        write_row.write_row(cursors, index_key, record_values, delta_count)
+                        write_row.write_row(cursors, index_key, record_values, delta_count as i64)
                     );
                     self.commit_state = AntijoinCommitState::PersistRCount {
                         deltas: std::mem::take(deltas),
@@ -818,7 +818,7 @@ impl IncrementalOperator for AntijoinOperator {
                     return_and_restore_if_io!(
                         &mut self.commit_state,
                         state,
-                        write_row.write_row(cursors, index_key, record_values, *l_weight)
+                        write_row.write_row(cursors, index_key, record_values, *l_weight as i64)
                     );
                     self.commit_state = AntijoinCommitState::PersistLIndex {
                         deltas: std::mem::take(deltas),
